// src/app/core/models/brewery.model.ts
export type Brewery = {
  id: string;
  name: string;
  brewery_type?: string | null;
  address_1?: string | null;
  address_2?: string | null;
  address_3?: string | null;
  city?: string | null;
  state_province?: string | null;
  postal_code?: string | null;
  country?: string | null;
  longitude?: number | null;
  latitude?: number | null;
  phone?: string | null;
  website_url?: string | null;
};